+++ CHANGELOG.md
@@ -11,10 +11,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - actuation: Extend ImuData with raw measurements
 - actuation: Log raw IMU measurements from pi3hat interface
+- envs: Added child environments UpkieServoPositions and UpkieServoTorques
 - envs: Start reward submodule with a wheeled inverted pendulum reward
+- envs: Wheeled inverted pendulum environment for reduced-model testing
 - utils: Add unit tests for Raspberry Pi utility functions
 - utils: Raise an exception when trying to configure agent from an interpreter
-- envs: Added child environments UpkieServoPositions and UpkieServoTorques 
 
 ### Changed