+++ CHANGELOG.md
@@ -26,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Avoid using `numpy.typing` for compatibility with older versions of NumPy
+- PID balancer: Fix dependencies for running on conda-less systems
 - deps: Update to palimpsest 2.2.0
 - envs: Refactor internal reward of `UpkieGroundVelocity` environment
 - envs: Use torque limits from model in `UpkieGroundVelocity` environment